requirements.txt
Django==2.2.24
django-leaflet==0.28.1
djangorestframework==3.12.4
